+++ R/logfile_-.R
@@ -1,7 +1,0 @@
-#' @export
-`logfile<-` <-
-function(x, value)
-{
-  UseMethod('logfile<-', x)
-}
-