+++ R/logformat.R
@@ -1,6 +1,0 @@
-#' @export
-`logformat` <-
-function(x)
-{
-  UseMethod('logformat', x)
-}