+++ tests/testthat/test-acceptance.R
@@ -17,21 +17,19 @@ test_that('Logger levels', {
   logfile(logger) <- file.path('base.log')
 
   level(logger) <- log4r:::DEBUG
-  expect_that(level(logger), equals(1))
+  expect_true(level(logger) == 1)
 
   level(logger) <- log4r:::INFO
-  expect_that(level(logger), equals(2))
+  expect_true(level(logger) == 2)
 
   level(logger) <- log4r:::WARN
-  expect_that(level(logger), equals(3))
+  expect_true(level(logger) == 3)
 
   level(logger) <- log4r:::ERROR
-  expect_that(level(logger), equals(4))
+  expect_true(level(logger) == 4)
 
   level(logger) <- log4r:::FATAL
-  expect_that(level(logger), equals(5))
-
-  level(logger) <- log4r:::DEBUG
+  expect_true(level(logger) == 5)
 
   unlink(logfile(logger))
 })
@@ -60,6 +58,32 @@ test_that('Creation of log file on first log entry', {
   expect_that(file.exists(logfile(logger)), is_true())
 
   unlink(logfile(logger))
+
+  fatal(logger, 'A Fatal Error Message')
+  expect_that(file.exists(logfile(logger)), is_true())
+
+  unlink(logfile(logger))
+})
+
+test_that('No creation of log file with insufficient level', {
+  logger <- create.logger()
+  logfile(logger) <- file.path('base.log')
+
+  level(logger) <- "INFO"
+  debug(logger, 'A Debugging Message')
+  expect_that(file.exists(logfile(logger)), is_false())
+
+  level(logger) <- "WARN"
+  info(logger, 'An Info Message')
+  expect_that(file.exists(logfile(logger)), is_false())
+
+  level(logger) <- "ERROR"
+  warn(logger, 'A Warning Message')
+  expect_that(file.exists(logfile(logger)), is_false())
+
+  level(logger) <- "FATAL"
+  error(logger, 'An Error Message')
+  expect_that(file.exists(logfile(logger)), is_false())
 
   fatal(logger, 'A Fatal Error Message')
   expect_that(file.exists(logfile(logger)), is_true())